src: Implement integrated or pop-out ConsoleView

+++ src/qml/ConsoleView.qml
@@ -6,11 +6,19 @@ import QtQuick.Effects
 Item {
     id: consoleView
 
-    readonly property int debuggerPaddingX : consoleView.parent != consoleViewLandingPad ? debuggerArea.width : 0
-    readonly property int headerPaddingY : consoleView.parent != consoleViewLandingPad && !root.landscapeMode ? root.headerItemHeight : 0
+    readonly property int debuggerPaddingX :
+        integratedMode ? 0 : parent != consoleViewDialogLandingPad ? debuggerArea.width : 0
+    readonly property int headerPaddingY :
+        integratedMode ? 0 : parent != consoleViewDialogLandingPad && !root.landscapeMode ? root.headerItemHeight : 0
 
-    x: ((parent.width - width - debuggerPaddingX) / 2)
-    y: visibility ? ((parent.height - height) / 2) + headerPaddingY : parent.height
+    property bool integratedMode : false
+    onIntegratedModeChanged: {
+        visibility = integratedMode
+    }
+    property bool expanded : false
+
+    x: integratedMode ? 0 : ((parent.width - width - debuggerPaddingX) / 2)
+    y: integratedMode ? 0 : visibility ? ((parent.height - height) / 2) + headerPaddingY : parent.height
     opacity: visibility ? opacityOverride : 0.0
     visible: opacity > 0.0
 
@@ -26,11 +34,17 @@ Item {
     readonly property bool modal : true
 
     function show() {
+        if (integratedMode)
+            expanded = true
+
         visibility = true;
         hideStdOut = false
     }
 
     function hide() {
+        if (integratedMode)
+            expanded = false
+
         visibility = false
     }
 
@@ -55,12 +69,14 @@ Item {
     }
 
     Behavior on y {
+        enabled: !integratedMode
         NumberAnimation {
             duration: dialogShadow.consoleAnimation
             easing.type: Easing.OutCubic
         }
     }
     Behavior on width {
+        enabled: !integratedMode
         NumberAnimation {
             duration: dialogShadow.consoleAnimation
             easing.type: Easing.OutCubic
@@ -77,7 +93,9 @@ Item {
         id: consoleRect
         anchors.fill: parent
         color: root.palette.base
-        radius: roundedCornersRadius
+        border.color: root.borderColor
+        border.width: 1
+        radius: roundedCornersRadiusMedium
         clip: true
 
         Column {
@@ -88,11 +106,13 @@ Item {
             Item {
                 id: consoleToolBar
                 width: parent.width
-                height: root.headerBarHeight
+                height: root.toolBarHeight
 
                 RowLayout {
                     anchors.fill: parent
                     TideToolButton {
+                        enabled: ((integratedMode && expanded) || !integratedMode) && consoleOutput.count > 0
+                        visible: enabled
                         icon.source: Qt.resolvedUrl("qrc:/assets/xmark.circle@2x.png")
                         icon.color: root.palette.button
                         leftPadding: paddingMedium
@@ -101,6 +121,8 @@ Item {
                         }
                     }
                     TideToolButton {
+                        enabled: ((integratedMode && expanded) || !integratedMode) && consoleOutput.count > 0
+                        visible: enabled
                         icon.source: !consoleView.hideStdOut ?
                                          Qt.resolvedUrl("qrc:/assets/line.3.horizontal.decrease.circle@2x.png")
                                        : Qt.resolvedUrl("qrc:/assets/line.3.horizontal.decrease.circle.fill@2x.png")
@@ -122,18 +144,42 @@ Item {
 
                     ToolButton {
                         text: qsTr("Stop")
+                        enabled: !integratedMode
+                        visible: !integratedMode && (wasmRunner.running || pyRunner.running)
                         font.bold: true
-                        visible: wasmRunner.running
                         onClicked: {
-                            wasmRunner.kill()
+                            if (wasmRunner)
+                                wasmRunner.kill()
+                            if (pyRunner)
+                                pyRunner.kill()
                         }
                     }
                     ToolButton {
                         text: qsTr("Hide")
+                        enabled: !integratedMode
+                        visible: !integratedMode
                         font.bold: true
                         rightPadding: paddingMedium
                         onClicked: {
                             consoleView.hide()
+                        }
+                    }
+                    ToolButton {
+                        Layout.alignment: Qt.AlignRight
+                        Layout.rightMargin: paddingMedium
+                        Layout.topMargin: expanded ? paddingSmall : 0
+                        visible: integratedMode
+                        enabled: integratedMode
+                        icon.source: expanded ?
+                                         Qt.resolvedUrl("qrc:/assets/chevron.compact.down@2x.png") :
+                                         Qt.resolvedUrl("qrc:/assets/chevron.compact.up@2x.png")
+                        icon.width: 32
+                        icon.height: 32
+                        Layout.preferredWidth: 32
+                        Layout.preferredHeight: 32
+                        font.pixelSize: 16
+                        onClicked: {
+                            expanded = !expanded
                         }
                     }
                 }
@@ -189,7 +235,7 @@ Item {
                 id: consoleInputField
                 font: fixedFont
                 width: parent.width
-                enabled: consoleView.inputEnabled
+                enabled: consoleView.inputEnabled && ((integratedMode && expanded) || !integratedMode)
                 height: enabled ? font.pixelSize + (paddingMid * 2) : 0
                 visible: height > 0
                 background: Item { }
@@ -223,7 +269,7 @@ Item {
             font: fixedFont
             text: !inputEnabled ? qsTr("Nothing running in the console yet.") :
                                   qsTr("Waiting for output...")
-            visible: consoleOutput.count === 0
+            visible: consoleOutput.count === 0 && (integratedMode && expanded)
             anchors.fill: parent
             horizontalAlignment: Label.AlignHCenter
             verticalAlignment: Label.AlignVCenter
@@ -237,6 +283,7 @@ Item {
         paddingRect: Qt.rect(0, 0, consoleRect.width, consoleRect.height)
         shadowBlur: 1.0
         shadowEnabled: true
+        visible: !integratedMode
     }
 }
